refactor(BaseballController.kt): move BaseballController.kt under baseball package, not domain
