+++ src/main/kotlin/baseball/Application.kt
@@ -1,7 +1,5 @@
 package baseball
 
-import baseball.domain.BaseballController
-
 fun main() {
     BaseballController().run()
 }
